backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
requests==2.31.0
redis==5.0.1
django-redis==5.4.0
pybreaker==1.0.1